Move src/{utils,components}/VisibilityCheckbox
