+++ frontend/src/components/VisibilityCheckbox.jsx
